Fix a couple of very minor bugs in serialice:
Correct the type of the parameter 'port' in the function
pnp_set_logical_device.  This should be a u16.  I'm not aware of this
causing a problem with any of the current mainboards as there ports
are all less than 256, but this will cause problems with mainboards
that have there PNP configuration port at 0x3F0.

Correct a comment in serialice.c.  The command "wi" is a write.

Adjust the code saving and restoring eax in start.S.  The code should
save and restore eax around the enqble of FPU and SSE.  If ENABLE_SSE
is defined the value of eax is trashed.  Again, this is mainly cosmetic
as eax is actually never used.

Signed-off-by: Mark Marshall <mark.marshall60@ntlworld.com>
Acked-by: Stefan Reinauer <stepan@coresystems.de>




git-svn-id: svn://svn.coreboot.org/serialice/trunk@28 8d503c2a-b2ab-11de-a960-3b3a9759a4e4