Add hooking mechanism to provide for special parsers
Right now the code hardcodes a lot of special handling.
The idea is to move these into functions that are run until
one is successful, or use a standard routine if none claimed
responsibility.

Later-on, these handlers could be chipset specific and enabled
automatically to provide a better view of what's going on.

Change-Id: I10cf9debd718cf5f60a652d071dcb9356ac79a03
Signed-off-by: Patrick Georgi <patrick.georgi@secunet.com>
Reviewed-on: http://review.coreboot.org/462
Tested-by: build bot (Jenkins)
Reviewed-by: Peter Stuge <peter@stuge.se>