Move SerialICE protocol support to separate file
serialice-com.c is responsible for the lowlevel details of
communication. serialice.c does the hooks and lua coordination.

No changes were made to functions operation. Two unused functions
are #if 0'd out, and there's some reordering going on to make gcc
happy, but any change inside a function is a bug.

Change-Id: Ic3a308dfd434d48c67018227707f8022c2e40050
Signed-off-by: Kyösti Mälkki <kyosti.malkki@gmail.com>
[pg: Reorganized Kyösti's patch set]
Signed-off-by: Patrick Georgi <patrick@georgi-clan.de>
Reviewed-on: http://review.coreboot.org/1066
Tested-by: build bot (Jenkins)