Only log Raw action lines from Qemu session
The only lines a qemu session with SerialICE needs to write in the
output are those produced with the core filters and marked with the
flag Raw "R...". Everything else can be replayed from these.

This is a principal change: the output of Qemu session is now completely
raw and one must pass the logfile thru replayer to have simplified view.

Change-Id: I6e4eebde41ca46072c545c02eabac848113bb5ee
Signed-off-by: Kyösti Mälkki <kyosti.malkki@gmail.com>
Reviewed-on: http://review.coreboot.org/2597
Tested-by: build bot (Jenkins)
Reviewed-by: Stefan Reinauer <stefan.reinauer@coreboot.org>