Add support for Intel D945GNT
This is based on the patch from Beata dalHagen, see
http://serialice.com/pipermail/serialice/2012-May/000372.html
This version has only been compile tested.

To share the code added by this and following patches of similar boards
the file originally added for the D945GCLF is renamed to be more generic.
Also, the Super I/O options get passed to the init function depending on
the configured board.

Change-Id: Ia4058d0f73f8357b97d7c0fe1868d9a289c4e2f7
Signed-off-by: Stefan Tauner <stefan.tauner@gmx.at>
Reviewed-on: http://review.coreboot.org/1578
Reviewed-by: Anton Kochkov <anton.kochkov@gmail.com>
Tested-by: build bot (Jenkins)
Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Reviewed-by: Kyösti Mälkki <kyosti.malkki@gmail.com>