EPIA-M850: Disable auto-reboot
The VX900 chipset on this board has a watchdog-like mechanism. This
reboots the board four seconds after it is powered on. It will not reboot
it a second time, but is still annoying when debugging.

Disable the timer that causes the reset. This prevents the board from
resetting 4 seconds after power on.

Change-Id: Ib604634ef4b33ddd3f552676ef5df7c4d142ead0
Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>
Reviewed-on: http://review.coreboot.org/3278
Tested-by: build bot (Jenkins)
Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Reviewed-by: Kyösti Mälkki <kyosti.malkki@gmail.com>