doc/*.md: Replace \302\240 sequence with standard space
Text within inline code blocks throughout the documentation contained
the \302\240 sequence (circumflex A, non-breaking space) in place of the
space character. Replace this sequence with a plain ASCII space.

This also fixes the first code block under the "Connect with GDB"
heading on Debugging.md, which was being rendered as normal text rather
than a code block.

The following command was used to make these replacements:

    for f in *.md; do sed -i "s/`echo -n '\0302\0240'`/ /g" $f; done

Change-Id: I292e137f6568691ea224f7532b1039e8ffc75376
Signed-off-by: Nicholas Chin <nic.c3.14@gmail.com>
Reviewed-on: https://review.coreboot.org/c/serialice/+/88579
Tested-by: Stefan Reinauer <stefan.reinauer@coreboot.org>
Reviewed-by: Stefan Reinauer <stefan.reinauer@coreboot.org>